+++ lib/kino_bumblebee/task_cell.ex
@@ -96,77 +96,6 @@ defmodule KinoBumblebee.TaskCell do
       label: "Text",
       tasks: [
         %{
-          id: "conversation",
-          label: "Conversation",
-          variants: [
-            %{
-              id: "blenderbot_400m",
-              label: "Blenderbot (400M parameters)",
-              docs_logo: "huggingface_logo.svg",
-              docs_url: "https://huggingface.co/facebook/blenderbot-400M-distill",
-              generation: %{
-                model_repo_id: "facebook/blenderbot-400M-distill",
-                tokenizer_repo_id: "facebook/blenderbot-400M-distill"
-              }
-            },
-            %{
-              id: "blenderbot_1b",
-              label: "Blenderbot (1B parameters)",
-              docs_logo: "huggingface_logo.svg",
-              docs_url: "https://huggingface.co/facebook/blenderbot-1B-distill",
-              generation: %{
-                model_repo_id: "facebook/blenderbot-1B-distill",
-                tokenizer_repo_id: "facebook/blenderbot-1B-distill"
-              }
-            },
-            %{
-              id: "blenderbot_3b",
-              label: "Blenderbot (3B parameters)",
-              docs_logo: "huggingface_logo.svg",
-              docs_url: "https://huggingface.co/facebook/blenderbot-3B",
-              generation: %{
-                model_repo_id: "facebook/blenderbot-3B",
-                tokenizer_repo_id: "facebook/blenderbot-3B"
-              }
-            },
-            %{
-              id: "dialogpt_small",
-              label: "DialoGPT (small)",
-              docs_logo: "huggingface_logo.svg",
-              docs_url: "https://huggingface.co/microsoft/DialoGPT-small",
-              generation: %{
-                model_repo_id: "microsoft/DialoGPT-small",
-                tokenizer_repo_id: "gpt2"
-              }
-            },
-            %{
-              id: "dialogpt_medium",
-              label: "DialoGPT (medium)",
-              docs_logo: "huggingface_logo.svg",
-              docs_url: "https://huggingface.co/microsoft/DialoGPT-medium",
-              generation: %{
-                model_repo_id: "microsoft/DialoGPT-medium",
-                tokenizer_repo_id: "gpt2"
-              }
-            },
-            %{
-              id: "dialogpt_large",
-              label: "DialoGPT (large)",
-              docs_logo: "huggingface_logo.svg",
-              docs_url: "https://huggingface.co/microsoft/DialoGPT-large",
-              generation: %{
-                model_repo_id: "microsoft/DialoGPT-large",
-                tokenizer_repo_id: "gpt2"
-              }
-            }
-          ],
-          params: [
-            %{field: "sequence_length", label: "Max input tokens", type: :number, default: 100},
-            %{field: "min_new_tokens", label: "Min new tokens", type: :number, default: nil},
-            %{field: "max_new_tokens", label: "Max new tokens", type: :number, default: 100}
-          ]
-        },
-        %{
           id: "fill_mask",
           label: "Fill-mask",
           variants: [
@@ -760,8 +689,7 @@ defmodule KinoBumblebee.TaskCell do
               label: "Number of images",
               type: :number,
               default: 2
-            },
-            %{field: "seed", label: "Seed", type: :number, default: nil}
+            }
           ],
           note:
             "this is a very involved task, the generation can take a long time if you run it on a CPU. To achieve a better quality increase the number of steps, 40 is usually a better default."
@@ -1361,55 +1289,6 @@ defmodule KinoBumblebee.TaskCell do
     ]
   end
 
-  defp to_quoted(%{"task_id" => "conversation"} = attrs) do
-    opts =
-      [compile: [batch_size: 1, sequence_length: attrs["sequence_length"]]] ++
-        maybe_defn_options(attrs)
-
-    generation_otps =
-      drop_nil_options(
-        min_new_tokens: attrs["min_new_tokens"],
-        max_new_tokens: attrs["max_new_tokens"]
-      )
-
-    %{generation: generation} = variant_from_attrs(attrs)
-
-    [
-      quote do
-        {:ok, model_info} = Bumblebee.load_model({:hf, unquote(generation.model_repo_id)})
-
-        {:ok, tokenizer} = Bumblebee.load_tokenizer({:hf, unquote(generation.tokenizer_repo_id)})
-
-        {:ok, generation_config} =
-          Bumblebee.load_generation_config({:hf, unquote(generation.model_repo_id)})
-
-        unquote_splicing(maybe_configure_generation(generation_otps))
-
-        serving =
-          Bumblebee.Text.conversation(model_info, tokenizer, generation_config, unquote(opts))
-      end,
-      quote do
-        frame = Kino.Frame.new()
-
-        inputs = [message: Kino.Input.text("Message")]
-        form = Kino.Control.form(inputs, submit: "Send message", reset_on_submit: [:message])
-
-        Kino.listen(form, nil, fn %{data: %{message: message}}, history ->
-          Kino.Frame.append(frame, Kino.Markdown.new("**Me:** #{message}"))
-
-          %{text: text, history: history} =
-            Nx.Serving.run(serving, %{text: message, history: history})
-
-          Kino.Frame.append(frame, Kino.Markdown.new("**Bot:** #{text}"))
-
-          {:cont, history}
-        end)
-
-        Kino.Layout.grid([frame, form], boxed: true, gap: 16)
-      end
-    ]
-  end
-
   defp to_quoted(
          %{"task_id" => "text_to_image", "variant_id" => "stable_diffusion_" <> _} = attrs
        ) do
@@ -1434,16 +1313,10 @@ defmodule KinoBumblebee.TaskCell do
 
         {:ok, clip} = Bumblebee.load_model({:hf, repository_id, subdir: "text_encoder"})
 
-        {:ok, unet} =
-          Bumblebee.load_model({:hf, repository_id, subdir: "unet"},
-            params_filename: "diffusion_pytorch_model.bin"
-          )
+        {:ok, unet} = Bumblebee.load_model({:hf, repository_id, subdir: "unet"})
 
         {:ok, vae} =
-          Bumblebee.load_model({:hf, repository_id, subdir: "vae"},
-            architecture: :decoder,
-            params_filename: "diffusion_pytorch_model.bin"
-          )
+          Bumblebee.load_model({:hf, repository_id, subdir: "vae"}, architecture: :decoder)
 
         {:ok, scheduler} = Bumblebee.load_scheduler({:hf, repository_id, subdir: "scheduler"})
 
@@ -1465,13 +1338,14 @@ defmodule KinoBumblebee.TaskCell do
       end,
       quote do
         text_input = Kino.Input.textarea("Text", default: unquote(generation.default_text))
-        form = Kino.Control.form([text: text_input], submit: "Run")
+        seed_input = Kino.Input.number("Seed")
+        form = Kino.Control.form([text: text_input, seed: seed_input], submit: "Run")
         frame = Kino.Frame.new()
 
-        Kino.listen(form, fn %{data: %{text: text}} ->
+        Kino.listen(form, fn %{data: %{text: text, seed: seed}} ->
           Kino.Frame.render(frame, Kino.Text.new("Running..."))
 
-          output = Nx.Serving.run(serving, text)
+          output = Nx.Serving.run(serving, %{prompt: text, seed: seed})
 
           for result <- output.results do
             Kino.Image.new(result.image)