+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased](https://github.com/livebook-dev/kino_bumblebee)
+
+### Added
+
+* Output streaming in the text generation task ([#32](https://github.com/livebook-dev/kino_bumblebee/pull/32))
+
+### Changed
+
+* Update Whisper model list
+* Moved seed to the prediction form in Stable Diffusion ([#41](https://github.com/livebook-dev/kino_bumblebee/pull/41))
+* Bumped `bumblebee` and `kino` dependency requirements
+
+### Removed
+
+* Removed the conversational task ([#41](https://github.com/livebook-dev/kino_bumblebee/pull/41))
+
 ## [v0.4.0](https://github.com/livebook-dev/kino_bumblebee/tree/v0.4.0) (2023-10-06)
 
 ### Added